refactor: move NamespaceName & related fn into mod
Moves the NamespaceName, associated errors and helper functions into
it's own module.
